11-legal-doc-summarizer/requirements.txt
requests
rich
click
PyPDF2
pyyaml
streamlit
python-dotenv
pytest
pytest-cov
